Initial reorganization for models and utils
