+++ cogs/models/weapon_list.py
@@ -1,5 +1,5 @@
-'''A useful module for computing weapon damage'''
-from troika_discord import util
+from cogs.utils.dice import roll_d6
+from cogs.models.weapon import Weapon
 
 
 def normalize_weapon_name(name):
@@ -9,56 +9,10 @@ def normalize_weapon_name(name):
 
 def normalize_armor_name(name):
     '''Converts armot name to lower case and removes " armor" from the back'''
-    return name.lower().replace(" armor", "")
+    return name.lower().replace(" armor", "").replace(" armour", "")
 
 
-ARMOR_OFFSETS = {
-    'light': 1,
-    'medium': 2,
-    'heavy': 3
-}
-
-
-class Weapon:
-    '''Defining the weapons as objects might be overkill, but I'd like to
-    support potential future options where DMs can support adding
-    new weapons via bot calls'''
-
-    def __init__(self, damage_table, name=None, style=None, two_handed=False, ignore_armor=False):
-        '''Defines a single weapon'''
-
-        if len(damage_table) != 7:
-            raise ValueError("You must provide a damage table of 7 elements for this weapon")
-
-        self.damage_table = damage_table
-        self.name = name
-        self.style = style
-        self.two_handed = two_handed
-        self.ignore_armor = ignore_armor
-
-    def roll_damage(self, armor_offset=0, damage_bonus=0):
-        '''Computes a damage roll, adjusting various modifiers'''
-        roll = util.roll_d6()
-        print("Roll ", roll)
-
-        if armor_offset > 0 and self.ignore_armor:
-            armor_offset -= 1
-
-        return max(roll - armor_offset + damage_bonus, 1)
-
-    def lookup_damage(self, roll):
-        '''Given a roll result, look up the damage from the weapon'''
-        if (roll < 1):
-            raise ValueError("The roll must be a positive value")
-
-        if (roll > 7):
-            roll = 7
-
-        # rolls range 1-7, arrays range 0-6
-        return self.damage_table[roll-1]
-
-
-class Weapons:
+class WeaponList:
     '''Represents a collection of weapons for the game'''
     def __init__(self):
         self.weapons = {}
@@ -73,6 +27,7 @@ class Weapons:
     def lookup_weapon(self, name):
         key = normalize_weapon_name(name)
         return self.weapons.get(key, None)
+
 
 # define standard weapons
 STANDARD_WEAPONS = [
@@ -110,6 +65,8 @@ STANDARD_WEAPONS = [
     Weapon([6, 8, 12, 16, 18, 24, 36], name="Dragon Fire", style="spell")
 ]
 
-all_weapons = Weapons()
+
+# Note that even though this is defined as a constant, you could still add new weapons to it
+ALL_WEAPONS = WeaponList()
 for w in STANDARD_WEAPONS:
-    all_weapons.add_weapon(w)
+    ALL_WEAPONS.add_weapon(w)
