+++ tests/conftest.py
@@ -1,4 +1,6 @@
+import pytest
 import sys
+import mock
 import os
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+#sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
